Fixed SQL table trait names
